Music Player Data - extend parsers There are now 2 tracks to parse tracks - default, & specific to LPs This also creates the default item metadata field `creator` for items not scanned by The Internet Archive
This is a simple, first step refactor where code is stable.  This can be improved upon & DRY'd up in the future
